vendor-local/src/django-tastytools/requirements.txt
django-tastypie
